Refactor chart service
