+++ src/app/game/charts/round-won-by/round-won-by.component.ts
@@ -1,8 +1,6 @@
 import {Component, Input, OnInit} from '@angular/core';
 import {ChartService} from '../chart.service';
 import {ChartDataSets, ChartType, RadialChartOptions} from 'chart.js';
-import {ChartData, WonWithMethods} from '../charts.types';
-import {concatAll, skip} from 'rxjs/operators';
 
 @Component({
   selector: 'app-round-won-by',
@@ -11,8 +9,7 @@ import {concatAll, skip} from 'rxjs/operators';
 })
 export class RoundWonByComponent implements OnInit {
 
-  @Input()
-  readonly gameId: number;
+  @Input() readonly gameId: number;
 
   readonly chartType: ChartType = 'radar';
   readonly chartLabels: string[] = ['EXPLOIT()', 'INFECT()', 'NOP()', 'OVERHEAR()', 'OVERLOAD()', 'PATCH()', 'SCAN()'];
@@ -22,22 +19,12 @@ export class RoundWonByComponent implements OnInit {
 
   chartData: ChartDataSets[] = [];
 
-  constructor(private readonly chartService: ChartService) {
+  constructor(private readonly service: ChartService) {
   }
 
   ngOnInit(): void {
-    this.chartService.getWonWithMethods(this.gameId)
-      .pipe(concatAll())
-      .pipe(skip(1))
-      .subscribe(bot => this.chartData.push(this.prepareChartData(bot))
-      );
-  }
-
-  private prepareChartData(bot: WonWithMethods): ChartData {
-    const result: ChartData = {data: [], label: ''};
-    result.label = bot.NAME;
-    Object.entries(bot.STATISTICS).forEach(([key, value]) => result.data.push(value));
-    return result;
+    this.service.getWonWithMethods(this.gameId)
+      .subscribe(bot => this.chartData.push(this.service.prepareWithMethodChartData(bot)));
   }
 
 }
